+++ NETCORE/src/Microsoft.ApplicationInsights.WorkerService/ApplicationInsightsExtensions.cs
@@ -175,6 +175,12 @@
                     {
                         exporterOptions.ConnectionString = serviceOptions.ConnectionString;
                     }
+                    
+                    // Copy credential to Azure Monitor Exporter
+                    if (serviceOptions.Credential != null)
+                    {
+                        exporterOptions.Credential = serviceOptions.Credential;
+                    }
 
                     if (!serviceOptions.EnableAdaptiveSampling)
                     {